+++ filament/src/PerViewUniforms.h
@@ -28,11 +28,14 @@
 
 #include <utils/EntityInstance.h>
 
+#include <random>
+
 namespace filament {
 
-struct FogOptions;
-struct DynamicResolutionOptions;
 struct AmbientOcclusionOptions;
+struct DynamicResolutionOptions;
+struct FogOptions;
+struct TemporalAntiAliasingOptions;
 struct VsmShadowOptions;
 
 struct CameraInfo;
@@ -48,15 +51,20 @@ class PerViewUniforms {
     using LightManagerInstance = utils::EntityInstance<LightManager>;
     using TextureHandle = backend::Handle<backend::HwTexture>;
 
-public:
-    PerViewUniforms(FEngine& engine) noexcept;
+    static constexpr uint32_t const SHADOW_SAMPLING_RUNTIME_PCF  = 0u;
+    static constexpr uint32_t const SHADOW_SAMPLING_RUNTIME_VSM  = 1u;
+    static constexpr uint32_t const SHADOW_SAMPLING_RUNTIME_DPCF = 2u;
 
-    void terminate(FEngine& engine);
+public:
+    explicit PerViewUniforms(FEngine& engine) noexcept;
+
+    void terminate(backend::DriverApi& driver);
 
     void prepareCamera(const CameraInfo& camera) noexcept;
     void prepareUpscaler(math::float2 scale, DynamicResolutionOptions const& options) noexcept;
     void prepareViewport(const filament::Viewport& viewport) noexcept;
-    void prepareTime(FEngine& engine, math::float4 const& userTime) noexcept;
+    void prepareTime(math::float4 const& userTime) noexcept;
+    void prepareTemporalNoise(TemporalAntiAliasingOptions const& options) noexcept;
     void prepareExposure(float ev100) noexcept;
     void prepareFog(const CameraInfo& camera, FogOptions const& options) noexcept;
     void prepareStructure(TextureHandle structure) noexcept;
@@ -75,6 +83,7 @@ public:
     // maybe these should have their own UBO, they're needed only when GENERATING the shadowmaps
     void prepareShadowVSM(TextureHandle texture, VsmShadowOptions const& options) noexcept;
     void prepareShadowPCF(TextureHandle texture) noexcept;
+    void prepareShadowDPCF(TextureHandle texture) noexcept;
 
     // update local data into GPU UBO
     void commit(backend::DriverApi& driver) noexcept;
@@ -91,6 +100,7 @@ private:
     backend::SamplerGroup mPerViewSb;
     backend::Handle<backend::HwSamplerGroup> mPerViewSbh;
     backend::Handle<backend::HwBufferObject> mPerViewUbh;
+    std::uniform_real_distribution<float> mUniformDistribution{0.0f, 1.0f};
 };
 
 } // namespace filament