+++ filament/src/RenderPass.cpp
@@ -286,7 +286,8 @@ void RenderPass::generateCommandsImpl(uint32_t extraFlags,
     materialVariant.setDirectionalLighting(renderFlags & HAS_DIRECTIONAL_LIGHT);
     materialVariant.setDynamicLighting(renderFlags & HAS_DYNAMIC_LIGHTING);
     materialVariant.setFog(renderFlags & HAS_FOG);
-    materialVariant.setVsm((renderFlags & HAS_VSM) && hasShadowing);
+    // DPCF uses the same color pass shader as VSM (but not for depth)
+    materialVariant.setVsm(((renderFlags & HAS_VSM) || (renderFlags & HAS_DPCF)) && hasShadowing);
     materialVariant.setShadowReceiver(false); // this is set per Renderable
 
     Command cmdColor;