+++ RELEASE_NOTES.md
@@ -3,7 +3,16 @@
 This file contains one line summaries of commits that are worthy of mentioning in release notes.
 A new header is inserted each time a *tag* is created.
 
-## v1.15.0 (currently main branch)
+## v1.15.2 (currently main branch)
+
+## v1.15.1
+
+- engine: add support for DPCF (PCF shadows with contact hardening).
+- engine: add support for Wayland and Vulkan.
+- engine: Fade lights out when close to light far plane.
+- Java: Add missing `Engine#destroySkinningBuffer` method.
+
+## v1.15.0
 
 - engine: Fix spotlights normal bias calculation [⚠️ **Material breakage**].
 - libimage: Fix loading spherical harmonics on certain locals.